+++ changelogs/unreleased/4422-fix-toolbar-allignment copy.yml
@@ -1,0 +1,9 @@
+description: Fix Service filtering when clicking on service relation
+issue-nr: 4099
+change-type: patch
+destination-branches:
+  - master
+  - iso5
+  - iso6
+sections: 
+  bugfix: "{{description}}"
